Fivedayforecast

+++ app/src/main/java/com/tronography/pixelweather/model/Main.kt
@@ -3,9 +3,9 @@ package com.tronography.pixelweather.model
 import com.google.gson.annotations.SerializedName
 
 class Main(
-        @SerializedName("temp_max") var tempMax: Double,
-        @SerializedName("pressure") var pressure: Double,
+        @SerializedName("temp_max") var tempMax: Float,
+        @SerializedName("pressure") var pressure: Float,
         @SerializedName("humidity") var humidity: Int,
-        @SerializedName("temp_min") var tempMin: Double,
-        @SerializedName("temp") var temp: Double
+        @SerializedName("temp_min") var tempMin: Float,
+        @SerializedName("temp") var temp: Float
 )
